readvar allow only x

+++ src/polynomial.cpp
@@ -4,7 +4,7 @@
 
 void Poly::print_polynomial(char x) {
   std::cout << x;
-  for (int i = 0; i < coefficients.size(); i++) {
+  for (size_t i = 0; i < coefficients.size(); i++) {
     std::cout << ' ' << coefficients[i];
   }
 }
